Switches to hooks (patch)
